Adjusted app version to match pushes. Moved BadgesRecyclerViewAdapter to proper package. Removed set text request from LeaderboardFragment.
