+++ app/src/main/java/com/example/workoutapp/ui/leaderboard/LeaderboardFragment.java
@@ -23,11 +23,9 @@ public class LeaderboardFragment extends Fragment {
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         leaderboardViewModel = new ViewModelProvider(requireActivity()).get(LeaderboardViewModel.class);
         View root = inflater.inflate(R.layout.fragment_leaderboard, container, false);
-        final TextView textView = root.findViewById(R.id.text_leaderboard);
         leaderboardViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
             @Override
             public void onChanged(@Nullable String s) {
-                textView.setText(s);
             }
         });
         return root;
